+++ src/main/java/racingcar/io/OutputManager.java
@@ -1,6 +1,7 @@
-package racingcar;
+package racingcar.io;
 
 import java.util.List;
+import racingcar.car.Car;
 
 public class OutputManager {
