+++ src/main/java/racingcar/name/Names.java
@@ -1,7 +1,6 @@
-package racingcar;
+package racingcar.name;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
