+++ src/test/java/racingcar/car/CarTest.java
@@ -1,10 +1,11 @@
-package racingcar;
+package racingcar.car;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.name.Name;
 
 class CarTest {
