+++ src/test/java/racingcar/io/InputManagerTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.io;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
