+++ src/test/java/racingcar/name/NamesTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.name;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
